pyproject.toml
[project]
name = "quant-krx300"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "finance-datareader>=0.9.100",
    "jinja2>=3.1.6",
    "networkx>=3.6.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]
